Added new logo and misc images; Added ADB receiver; Updated the build.sh to allow device installation and updated README.md

+++ build.sh
@@ -7,14 +7,12 @@
 # On gLinux, use 'export JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64'
 
 #-----------------------------------------------------------------------------
-INSTALL=false
 DEVICE=""
 PIPELINE=false
-USAGE="./build.sh [-i] [-s | --device SERIAL_NUMBER]\n\ti\t: install to phone\n\t-s\t: android serial number, if there is more than one device"
+USAGE="./build.sh [[-s | --device] SERIAL_NUMBER]"
 while [[ "$#" -gt 0 ]]; do
     case $1 in
-        -i) INSTALL=true; ;;
-        -s|--device) DEVICE="-s $2"; shift ;;
+        -s|--device) DEVICE="$2"; shift ;;
         -p) PIPELINE=true; shift ;;
         -h|--help) echo $USAGE; exit 0 ;;
         *) echo "Unknown parameter passed: $1"; exit 1 ;;
@@ -119,20 +117,23 @@ log
 GRADLE_ZIP_REMOTE_FILE=gradle-${GRADLE_DOWNLOAD_VERSION}-bin.zip
 GRADLE_ZIP_DEST_PATH=~/Desktop/${GRADLE_DOWNLOAD_VERSION}.zip
 
-if [[ "$PIPELINE" = true ]]; then
-  log "Download gradle binary from the web ${GRADLE_ZIP_REMOTE_FILE} to ${GRADLE_ZIP_DEST_PATH} using wget"
-  wget -O ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}
-  log
-else
-  log "Download gradle binary from the web ${GRADLE_ZIP_REMOTE_FILE} to ${GRADLE_ZIP_DEST_PATH} using curl"
-  COMMAND="curl -L -o ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}"
-  sudo curl -L -o ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}
+if [[ ! -f "$GRADLE_ZIP_DEST_PATH" ]]; then
+  log "--> Downloading GRADLE"
+  if [[ "$PIPELINE" = true ]]; then
+    log "Download gradle binary from the web ${GRADLE_ZIP_REMOTE_FILE} to ${GRADLE_ZIP_DEST_PATH} using wget"
+    wget -O ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}
+    log
+  else
+    log "Download gradle binary from the web ${GRADLE_ZIP_REMOTE_FILE} to ${GRADLE_ZIP_DEST_PATH} using curl"
+    COMMAND="curl -L -o ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}"
+    sudo curl -L -o ${GRADLE_ZIP_DEST_PATH} https://services.gradle.org/distributions/${GRADLE_ZIP_REMOTE_FILE}
+  fi
 fi
 
 
 GRADLE_UNZIP_HOSTING_FOLDER=/opt/gradle-${GRADLE_DOWNLOAD_VERSION}
 log "Unzip gradle zipfile ${GRADLE_ZIP_DEST_PATH} to ${GRADLE_UNZIP_HOSTING_FOLDER}"
-sudo unzip -n -d ${GRADLE_UNZIP_HOSTING_FOLDER} ${GRADLE_ZIP_DEST_PATH}
+unzip -n -d ${GRADLE_UNZIP_HOSTING_FOLDER} ${GRADLE_ZIP_DEST_PATH}
 log
 
 
@@ -182,5 +183,9 @@ if [[ $BUILD_EXIT_CODE -eq 0 ]]; then
   log
 fi
 
+if [[ ! -z $DEVICE ]]; then
+  log "installing on $DEVICE"
+  adb -s $DEVICE install ./build/outputs/apk/phone/debug/talkback-phone-debug.apk
+fi
 
 exit $BUILD_EXIT_CODE   ### This should be the last line in this file
